This commit is big, because several important changes:
The find dialog comes to be usable with some troubles;
the enableFindButton() slot and the find() signal didn't work.

I add some new functionalities to the dialog eg.: search backward
and find whole words. Thus, the translations will be change in a
feature commit.

The ommitted ned_hu.qm added to translations/ so there is no more
error messages, which claims for this file.